ADM-949 [backend]: exclude the weekend when we calculate pr lead time and dev mean time to recovery (#1462)
* ADM-949 [backend]: finish logic

* ADM-949 [backend]: add log

* ADM-949 [frontend]: fix e2e test

* ADM-949 [frontend]: fix checkstyle

* ADM-949 [backend]: fix sonar

* ADM-949 [backend]: fix sonar

* ADM-949 [backend]: refactor

* ADM-949 [backend]: fix bug

+++ backend/src/main/java/heartbeat/service/report/calculator/DeploymentFrequencyCalculator.java
@@ -26,7 +26,7 @@ public class DeploymentFrequencyCalculator {
 	private final WorkDay workDay;
 
 	public DeploymentFrequency calculate(List<DeployTimes> deployTimes, Long startTime, Long endTime) {
-		int timePeriod = workDay.calculateWorkDaysBetween(startTime, endTime);
+		long timePeriod = workDay.calculateWorkDaysBetween(startTime, endTime);
 
 		List<DeploymentFrequencyOfPipeline> deploymentFrequencyOfPipelines = deployTimes.stream().map((item) -> {
 			int passedDeployInfosCount = (int) item.getPassed()
